Rename Delay behavioral files
